backend/requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.1.7
djangorestframework==3.16.0
drf-yasg==1.21.10
django-cors-headers==4.6.0
django-filter==23.5
fonttools==4.42.1
fpdf2==2.7.5
inflection==0.5.1
numpy==1.25.2
packaging==24.2
pandas==2.1.0
Pillow==10.0.0
pypdf==3.16.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
django-environ==0.11.2
PyMySQL==1.1.0
whitenoise==6.7.0
psutil==6.0.0
dj-database-url==2.2.0
django-csp==3.8
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
qrcode==7.4.2
